Add files via upload
Interactive drawing file. Draw a 1 or 0 on a 5x5 canvas on tkinter (left-click = paint, right-click = erase). Then click whether it is a 1 or 0. This is just to set the actual label on what you just drew. Click the save button, and then click test. The neural network's answer will appear on the python shell. Small fix to test_convtemp function.